appdev: move CMakePresets.json to container and invert include direction
The toolchain definitions  are part of the container, so the CMakePresets.json should also be part of the container.
This also drops the include of app_presets.json. The include should now be done the other way around, i.e.
 - remove the old symlink to CMakePresets.json
 - rename app_presets.json to CMakePresets.json
 - add "include": [ "/build/cmake/BasePresets.json"]
 - add "inherits": "base" to the configurePreset

This is a more natural include direction and allows adding addition toolchain preset files,
that can be referenced in apps using the appropriate include file